Build in windows and linux CLang compiling has a problem in windows

+++ src/app/pars-park-app/model/settings/comport-model.hpp
@@ -18,7 +18,7 @@ namespace anar::parspark::model {
       }
 
       std::string PortName;
-      uint16_t BaudRate{19200};
+      uint32_t BaudRate{19200};
    };
 }  // namespace anar::parspark::model
 #endif  // ANAR_PARS_PARK_MODEL_COMPORT_HPP
